chore: refactor interrupts
Similar to Buffer.isBuffer or Stream.isStream, interrupts
have now `Interrupt.isInterrupt` to check if an object is an interrupt.

Every derived class has the isInterrupt static method
to check if an object is an interrupt of that particular type.

Eg:

```js
// true if interrupt is a FederatedConnectionInterrupt
FederatedConnectionInterrupt.isInterrupt(interrupt)
```

This is works both in javascript for matching the object returning true
or false but also is a user-defined type guard in typescript which gives.